+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/AllJavaAstTreeDumpTest.java
@@ -15,7 +15,6 @@ import org.junit.platform.suite.api.Suite;
     Java15TreeDumpTest.class,
     Java16TreeDumpTest.class,
     Java17TreeDumpTest.class,
-    Java20PreviewTreeDumpTest.class,
     Java21TreeDumpTest.class,
     Java21PreviewTreeDumpTest.class
 })