+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/ast/KotlinTestingDsl.kt
@@ -34,7 +34,7 @@ enum class JavaVersion : Comparable<JavaVersion> {
     J17,
     J18,
     J19,
-    J20, J20__PREVIEW,
+    J20,
     J21, J21__PREVIEW,
     J22, J22__PREVIEW;